chore(refactor test codes)
