+++ package.json
@@ -5,7 +5,7 @@
   "main": "server.js",
   "scripts": {
     "start": "nodemon --exec babel-node src/server.js",
-    "test": "nyc mocha --require @babel/register --require @babel/polyfill src/test --exit",
+    "test": "nyc mocha --require @babel/register --require @babel/polyfill src/test",
     "coverage": "nyc report --reporter=text-lcov | coveralls"
   },
   "repository": {
