+++ src/test/article.test.js
@@ -4,138 +4,131 @@ import chaiHttp from 'chai-http';
 import app from '../server';
 
 // token
-import token from './register.test';
+// import token from './register.test';
 
 chai.use(chaiHttp);
 chai.should();
 
-describe.skip('articles', () => {
+describe('articles', () => {
     describe('create article', () => {
         // error om wrong token
-        describe('POST create article', () => {
-            it('should give error on wrong token', (done) => {
-                chai.request(app)
-                    .post('/api/v1/articles')
-                    .set('Authorization', `bearer wrong token`)
-                    .send({
-                        title: 'my first official project',
-                        article: 'i was an intern at Hotels.ng.',
-                        authorId: 1
-                    })
-                    .end((err, res) => {
-                        res.should.have.status(403)
-                    })
-                done();
-            })
+        it('should give error on wrong token', (done) => {
+            chai.request(app)
+                .post('/api/v1/articles')
+                .set('Authorization', `bearer wrong token`)
+                .send({
+                    title: 'my first official project',
+                    article: 'i was an intern at Hotels.ng.',
+                    authorId: 1
+                })
+                .end((err, res) => {
+                    res.should.have.status(403);
+                    res.body.should.be.a('object');
+                })
+            done();
         })
 
 
         // error on empty body values
-        describe('POST create article', () => {
-            it('should give error on empty body values', (done) => {
-                chai.request(app)
-                    .post('/api/v1/articles')
-                    .set('Authorization', `bearer wrong token`)
-                    .send({
-                        title: 'my first official project',
-                        article: '',
-                        authorId: 1
-                    })
-                    .end((err, res) => {
-                        res.should.have.status(400)
-                    })
-                done();
-            })
+        it('should give error on empty body values', (done) => {
+            chai.request(app)
+                .post('/api/v1/articles')
+                .set('Authorization', `bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6Im1vaGFtbWVkIiwicGFzc3dvcmQiOiJpYnJhaGltIiwiaWF0IjoxNTcyODc1ODk2LCJleHAiOjE1NzI5NjIyOTZ9.Xd1C7c4T4DrQY5hZV0-J7dNmH1ioqwuH-956yDLMLlE`)
+                .send({
+                    title: 'my first official project',
+                    article: '',
+                    authorId: 1
+                })
+                .end((err, res) => {
+                    res.should.have.status(400);
+                    res.body.should.be.a('object');
+                })
+            done();
         })
 
-        describe('POST create article', () => {
-            it('should create a new article', (done) => {
-                chai.request(app)
-                    .post('/api/v1/articles')
-                    .set('Authorization', `bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6Im1vaGFtbWVkIiwicGFzc3dvcmQiOiJpYnJhaGltIiwiaWF0IjoxNTcyNzUyNjE1LCJleHAiOjE1NzI4MzkwMTV9.aNUBP0nNhk5etK-Fb98UzDQOZH1sPIrnbFsAEsRiAVo`)
-                    .send({
-                        title: 'my first official project',
-                        article: 'i was an intern at Hotels.ng.',
-                        authorId: 1
-                    })
-                    .end((err, res) => {
-                        res.should.have.status(201)
-                    })
-                done();
-            })
+
+        it('should create a new article', (done) => {
+            chai.request(app)
+                .post('/api/v1/articles')
+                .set('Authorization', `bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6Im1vaGFtbWVkIiwicGFzc3dvcmQiOiJpYnJhaGltIiwiaWF0IjoxNTcyODc1ODk2LCJleHAiOjE1NzI5NjIyOTZ9.Xd1C7c4T4DrQY5hZV0-J7dNmH1ioqwuH-956yDLMLlE`)
+                .send({
+                    title: 'my first official project',
+                    article: 'i was an intern at Hotels.ng.',
+                    authorId: 1
+                })
+                .end((err, res) => {
+                    res.should.have.status(201);
+                    res.body.should.be.a('object');
+                })
+            done();
         })
+
     })
 
 
 
     // test modify route
-    describe.skip('modify article', () => {
+    describe('PATCH modify article', () => {
         // give error on wrong token
-        describe('PATCH modify article', () => {
-            const id = 3;
-            it('should give an error on wrong token', (done) => {
-                chai.request(app)
-                    .patch(`/api/v1/articles/${id}`)
-                    .set('Authorization', `bearer wrong token`)
-                    .send({
-                        title: 'vacation',
-                        article: 'i travlled to my home town during the school holiday'
-                    })
-                    .end((err, res) => {
-                        res.should.have.status(403)
-                    })
-                    done();
-            })
+        const id = 10;
+        it('should give an error on wrong token', (done) => {
+            chai.request(app)
+                .patch(`/api/v1/articles/${id}`)
+                .set('Authorization', `bearer wrong token`)
+                .send({
+                    title: 'vacation',
+                    article: 'i travlled to my home town during the school holiday'
+                })
+                .end((err, res) => {
+                    res.should.have.status(403);
+                    res.body.should.be.a('object');
+                })
+            done();
         })
 
         // modify an article
-        describe('PATCH modify article', () => {
-            const id = 8;
-            it('should modify an article', (done) => {
-                chai.request(app)
-                    .patch(`/api/v1/articles/${id}`)
-                    .set('Authorization', `bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6Im1vaGFtbWVkIiwicGFzc3dvcmQiOiJpYnJhaGltIiwiaWF0IjoxNTcyNzUyNjE1LCJleHAiOjE1NzI4MzkwMTV9.aNUBP0nNhk5etK-Fb98UzDQOZH1sPIrnbFsAEsRiAVo`)
-                    .send({
-                        title: 'vacation',
-                        article: 'i travlled to my home town during the school holiday'
-                    })
-                    .end((err, res) => {
-                        res.should.have.status(200)
-                    })
-                    done();
-            })
+        it('should modify an article', (done) => {
+            chai.request(app)
+                .patch(`/api/v1/articles/${id}`)
+                .set('Authorization', `bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6Im1vaGFtbWVkIiwicGFzc3dvcmQiOiJpYnJhaGltIiwiaWF0IjoxNTcyODc1ODk2LCJleHAiOjE1NzI5NjIyOTZ9.Xd1C7c4T4DrQY5hZV0-J7dNmH1ioqwuH-956yDLMLlE`)
+                .send({
+                    title: 'vacation',
+                    article: 'i travlled to my home town during the school holiday'
+                })
+                .end((err, res) => {
+                    res.should.have.status(200);
+                    res.body.should.be.a('object');
+                })
+            done();
         })
     })
 
 
     // test delete route
-    describe.skip('delete article', () => {
+    describe('delete article', () => {
         // error on wrong token
-        describe('DELETE article', () => {
-            const id = 4;
-            it('should delete an article', (done) => {
-                chai.request(app)
+        const id = 4;
+        it('should delete an article', (done) => {
+            chai.request(app)
                 .delete(`/api/v1/articles/${id}`)
                 .set('Authorization', `bearer wrong token`)
                 .end((err, res) => {
-                    res.should.have.status(403)
+                    res.should.have.status(403);
+                    res.body.should.be.a('object');
                 })
-                done();
-            })
+            done();
         })
 
         // delete an article
-        describe('DELETE article', () => {
-            const id = 5;
-            it('should delete an article', (done) => {
-                chai.request(app)
+        it('should delete an article', (done) => {
+            chai.request(app)
                 .delete(`/api/v1/articles/${id}`)
-                .set('Authorization', `bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6Im1vaGFtbWVkIiwicGFzc3dvcmQiOiJpYnJhaGltIiwiaWF0IjoxNTcyNzUyNjE1LCJleHAiOjE1NzI4MzkwMTV9.aNUBP0nNhk5etK-Fb98UzDQOZH1sPIrnbFsAEsRiAVo`)
+                .set('Authorization', `bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6Im1vaGFtbWVkIiwicGFzc3dvcmQiOiJpYnJhaGltIiwiaWF0IjoxNTcyODc1ODk2LCJleHAiOjE1NzI5NjIyOTZ9.Xd1C7c4T4DrQY5hZV0-J7dNmH1ioqwuH-956yDLMLlE`)
                 .end((err, res) => {
-                    res.should.have.status(200)
+                    res.should.have.status(200);
+                    res.body.should.be.a('object');
                 })
-                done();
-            })
+            done();
         })
     })
