+++ src/test/comment.test.js
@@ -3,21 +3,21 @@ import chaiHttp from 'chai-http';
 
 import app from '../server';
 
-import token from './register.test';
+// import token from './register.test';
 
 chai.use(chaiHttp);
 chai.should();
 
-describe.skip('POST comment', () => {
+describe('POST comment', () => {
     // test article comment
-    describe('post article comment', () => {
-        const id = 5;
+    describe('POST article comment', () => {
+        const id = 6;
 
         // error on empty body values
         it('should give an error on empty body values', (done) => {
             chai.request(app)
                 .post(`api/v1/articles/${id}/comment`)
-                .set('Authorization', `bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6Im1vaGFtbWVkIiwicGFzc3dvcmQiOiJpYnJhaGltIiwiaWF0IjoxNTcyNzUyNjE1LCJleHAiOjE1NzI4MzkwMTV9.aNUBP0nNhk5etK-Fb98UzDQOZH1sPIrnbFsAEsRiAVo`)
+                .set('Authorization', `bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6Im1vaGFtbWVkIiwicGFzc3dvcmQiOiJpYnJhaGltIiwiaWF0IjoxNTcyODc1ODk2LCJleHAiOjE1NzI5NjIyOTZ9.Xd1C7c4T4DrQY5hZV0-J7dNmH1ioqwuH-956yDLMLlE`)
                 .send({
                     comment: '',
                     authorId: 1
@@ -49,7 +49,7 @@ describe.skip('POST comment', () => {
         it('should post an article comment', (done) => {
             chai.request(app)
                 .post(`api/v1/articles/${id}/comment`)
-                .set('Authorization', `bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6Im1vaGFtbWVkIiwicGFzc3dvcmQiOiJpYnJhaGltIiwiaWF0IjoxNTcyNzUyNjE1LCJleHAiOjE1NzI4MzkwMTV9.aNUBP0nNhk5etK-Fb98UzDQOZH1sPIrnbFsAEsRiAVo`)
+                .set('Authorization', `bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6Im1vaGFtbWVkIiwicGFzc3dvcmQiOiJpYnJhaGltIiwiaWF0IjoxNTcyODc1ODk2LCJleHAiOjE1NzI5NjIyOTZ9.Xd1C7c4T4DrQY5hZV0-J7dNmH1ioqwuH-956yDLMLlE`)
                 .send({
                     comment: 'nice article.',
                     authorId: 1
@@ -66,12 +66,12 @@ describe.skip('POST comment', () => {
     // test gif comment
     describe('POST gif comment', () => {
         const id = 7;
-        
+
         // error on empty body values
         it('should give an error on empty body values', (done) => {
             chai.request(app)
                 .post(`/api/v1/gifs/${id}/comment`)
-                .set('Authorization', `bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6Im1vaGFtbWVkIiwicGFzc3dvcmQiOiJpYnJhaGltIiwiaWF0IjoxNTcyNzUyNjE1LCJleHAiOjE1NzI4MzkwMTV9.aNUBP0nNhk5etK-Fb98UzDQOZH1sPIrnbFsAEsRiAVo`)
+                .set('Authorization', `bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6Im1vaGFtbWVkIiwicGFzc3dvcmQiOiJpYnJhaGltIiwiaWF0IjoxNTcyODc1ODk2LCJleHAiOjE1NzI5NjIyOTZ9.Xd1C7c4T4DrQY5hZV0-J7dNmH1ioqwuH-956yDLMLlE`)
                 .send({
                     comment: '',
                     authorId: 1
@@ -80,7 +80,7 @@ describe.skip('POST comment', () => {
                     res.should.have.status(400);
                     res.body.should.be.a('object');
                 })
-                done();
+            done();
         })
 
         // error on wromg token
@@ -96,7 +96,7 @@ describe.skip('POST comment', () => {
                     res.should.have.status(403);
                     res.body.should.be.a('object');
                 })
-                done();
+            done();
         })
 
 
@@ -104,7 +104,7 @@ describe.skip('POST comment', () => {
         it('should post a gif comment', (done) => {
             chai.request(app)
                 .post(`/api/v1/gifs/${id}/comment`)
-                .set('Authorization', `bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6Im1vaGFtbWVkIiwicGFzc3dvcmQiOiJpYnJhaGltIiwiaWF0IjoxNTcyNzUyNjE1LCJleHAiOjE1NzI4MzkwMTV9.aNUBP0nNhk5etK-Fb98UzDQOZH1sPIrnbFsAEsRiAVo`)
+                .set('Authorization', `bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6Im1vaGFtbWVkIiwicGFzc3dvcmQiOiJpYnJhaGltIiwiaWF0IjoxNTcyODc1ODk2LCJleHAiOjE1NzI5NjIyOTZ9.Xd1C7c4T4DrQY5hZV0-J7dNmH1ioqwuH-956yDLMLlE`)
                 .send({
                     comment: 'what a funny gif.',
                     authorId: 1
@@ -113,7 +113,7 @@ describe.skip('POST comment', () => {
                     res.should.have.status(201);
                     res.body.should.be.a('object');
                 })
-                done();
+            done();
         })
     })
 })
