+++ src/test/get.test.js
@@ -7,7 +7,7 @@ chai.use(chaiHttp);
 chai.should();
 
 describe('GET', () => {
-    describe.skip('GET feed', () => {
+    describe('GET feed', () => {
         // error on wrong token
         it('should give an error on wrong token', (done) => {
             chai.request(app)
@@ -21,23 +21,23 @@ describe('GET', () => {
         })
 
         // no feed available
-        it('should give an error if there are no feeds', (done) => {
-            chai.request(app)
-            .get('/api/v1/feed')
-            .set('Authorization', `bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6Im1vaGFtbWVkIiwicGFzc3dvcmQiOiJpYnJhaGltIiwiaWF0IjoxNTcyNzUyNjE1LCJleHAiOjE1NzI4MzkwMTV9.aNUBP0nNhk5etK-Fb98UzDQOZH1sPIrnbFsAEsRiAVo`)
-            .end((err, res) => {
-                res.should.have.status(400);
-                res.body.should.be.a('object');
-            })
-            done();
-        })
+        // it('should give an error if there are no feeds', (done) => {
+        //     chai.request(app)
+        //     .get('/api/v1/feed')
+        //     .set('Authorization', `bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6Im1vaGFtbWVkIiwicGFzc3dvcmQiOiJpYnJhaGltIiwiaWF0IjoxNTcyODc1ODk2LCJleHAiOjE1NzI5NjIyOTZ9.Xd1C7c4T4DrQY5hZV0-J7dNmH1ioqwuH-956yDLMLlE`)
+        //     .end((err, res) => {
+        //         res.should.have.status(400);
+        //         res.body.should.be.a('object');
+        //     })
+        //     done();
+        // })
 
 
         // get all feeds
         it('should get all articles and gif', (done) => {
             chai.request(app)
             .get('/api/v1/feed')
-            .set('Authorization', `bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6Im1vaGFtbWVkIiwicGFzc3dvcmQiOiJpYnJhaGltIiwiaWF0IjoxNTcyNzUyNjE1LCJleHAiOjE1NzI4MzkwMTV9.aNUBP0nNhk5etK-Fb98UzDQOZH1sPIrnbFsAEsRiAVo`)
+            .set('Authorization', `bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6Im1vaGFtbWVkIiwicGFzc3dvcmQiOiJpYnJhaGltIiwiaWF0IjoxNTcyODc1ODk2LCJleHAiOjE1NzI5NjIyOTZ9.Xd1C7c4T4DrQY5hZV0-J7dNmH1ioqwuH-956yDLMLlE`)
             .end((err, res) => {
                 res.should.have.status(200);
                 res.body.should.be.a('object');
