+++ src/test/register.test.js
@@ -7,7 +7,7 @@ chai.use(chaiHttp);
 chai.should();
 
 // let signUpToken = null;
-let token = 'token';
+// let token = '';
 
 describe('POST register', () => {
     // error if fields are empty
@@ -26,44 +26,43 @@ describe('POST register', () => {
                     address: '4, alomosho'
                 })
                 .end((err, res) => {
-                    res.should.have.status(400)
+                    res.should.have.status(400);
+                    res.body.should.be.a('object');
                 })
             done();
         })
-    })
 
 
-    // error if a user already exists
-    describe('POST sign up', () => {
+        // error if a user already exists
         it('should give error is a user already exists', (done) => {
             chai.request(app)
                 .post('/api/v1/auth/create-user')
                 .send({
                     firstName: 'mama',
                     lastName: 'omo',
-                    email: 'new usermkmkm',
-                    password: 'admin',
+                    email: 'mohammed',
+                    password: 'ibrahim',
                     gender: 'male',
                     jobRole: 'assistant',
                     department: 'engineer',
                     address: '4, alomosho'
                 })
                 .end((err, res) => {
-                    res.should.have.status(400)
+                    res.should.have.status(400);
+                    res.body.should.be.a('object');
                 })
             done();
         })
-    })
 
-    // register new user
-    describe.skip('POST sign up', () => {
+
+        // register new user
         it('should sign up a user', (done) => {
             chai.request(app)
                 .post('/api/v1/auth/create-user')
                 .send({
                     firstName: 'mohammed',
                     lastName: 'ibrahim',
-                    email: 'new',
+                    email: 'damilola',
                     password: 'admin',
                     gender: 'male',
                     jobRole: 'assistant',
@@ -71,81 +70,80 @@ describe('POST register', () => {
                     address: '4, alomosho'
                 })
                 .end((err, res) => {
-                    res.should.have.status(201)
-                    // signUpToken = res.body.data.token;
+                    res.should.have.status(201);
+                    res.body.should.be.a('object');
                 })
             done();
         })
     })
 
-    //
+
     // test log in
     // error if fields are empty
-    describe('POST sign up', () => {
+    describe('POST login', () => {
         it('should give error if fields are empty', (done) => {
             chai.request(app)
-                .post('/api/v1/auth/create-user')
+                .post('/api/v1/auth/signin')
                 .send({
                     email: 'new usermkmkm',
                     password: ''
                 })
                 .end((err, res) => {
-                    res.should.have.status(400)
+                    res.should.have.status(400);
+                    res.body.should.be.a('object');
                 })
             done();
         })
-    })
 
-    // error if user does not exist
-    describe('POST login', () => {
+
+        // error if user does not exist
         it('should give error if user does not exist', (done) => {
             chai.request(app)
-            .post('/api/v1/auth/signin')
-            .send({
-                email: 'non existing user',
-                password: 'password'
-            })
-            .end((err, res) => {
-                res.should.have.status(400)
-            })
+                .post('/api/v1/auth/signin')
+                .send({
+                    email: 'non existing user',
+                    password: 'password'
+                })
+                .end((err, res) => {
+                    res.should.have.status(400)
+                    res.body.should.be.a('object');
+                })
             done();
         })
-    })
 
-    // login a user
-    describe('POST login', () => {
+
+        // login a user
         it('should log in an existing user', (done) => {
             chai.request(app)
-            .post('/api/v1/auth/signin')
-            .send({
-                email: 'mohammed',
-                password: 'ibrahim'
-            })
-            .end((err, res) => {
-                res.should.have.status(201)
-                token = res.body.data.token
-                console.log(token)
-            })
+                .post('/api/v1/auth/signin')
+                .send({
+                    email: 'mohammed',
+                    password: 'ibrahim'
+                })
+                .end((err, res) => {
+                    res.should.have.status(201);
+                    res.body.should.be.a('object');
+                    // token = res.body.data.token
+                })
             done();
         })
-    })
 
 
-    // error for incorrect email or password
-    describe('POST login', () => {
+        // error for incorrect email or password
         it('should give error for incorrect email or password', (done) => {
             chai.request(app)
-            .post('/api/v1/auth/signin')
-            .send({
-                email: 'new',
-                password: 'bypass'
-            })
-            .end((err, res) => {
-                res.should.have.status(403)
-            })
+                .post('/api/v1/auth/signin')
+                .send({
+                    email: 'new',
+                    password: 'bypass'
+                })
+                .end((err, res) => {
+                    res.should.have.status(403);
+                    res.body.should.be.a('object');
+                })
             done();
         })
     })
 })
 
-export default token;
+// export default token;
